+++ src/collective/cover/behaviors/interfaces.py
@@ -1,6 +1,7 @@
 # coding: utf-8
 from collective.cover import _
 from plone.directives import form
+from plone.namedfile.field import NamedBlobImage
 from plone.supermodel import model
 from zope import schema
 from zope.interface import alsoProvides
@@ -32,3 +33,19 @@ alsoProvides(IRefresh, form.IFormFieldProvider)
 def validate_ttl(value):
     if value <= 0:
         raise Invalid(_(u'Value must be greater than zero.'))
+
+
+class IBackgroundImage(model.Schema):
+
+    """Adds the possibility of having a background image on the item."""
+
+    background_image = NamedBlobImage(
+        title=_(u'Background image'),
+        description=_(
+            u'Sets the background image to be used on the item. '
+            u'For accessibility reasons, you should not use background images as the sole method of conveying important information.',
+        ),
+        required=False,
+    )
+
+alsoProvides(IBackgroundImage, form.IFormFieldProvider)