+++ src/collective/cover/tests/test_background_image_behavior.py
@@ -1,0 +1,60 @@
+# -*- coding: utf-8 -*-
+from collective.cover.behaviors.interfaces import IBackgroundImage
+from collective.cover.testing import INTEGRATION_TESTING
+from plone import api
+from plone.behavior.interfaces import IBehavior
+from plone.dexterity.interfaces import IDexterityFTI
+from plone.dexterity.schema import SchemaInvalidatedEvent
+from zope.component import queryUtility
+from zope.event import notify
+
+import unittest
+
+
+class BackgroundImageTestCase(unittest.TestCase):
+
+    layer = INTEGRATION_TESTING
+
+    def _enable_background_image_behavior(self):
+        fti = queryUtility(IDexterityFTI, name='collective.cover.content')
+        behaviors = list(fti.behaviors)
+        behaviors.append(IBackgroundImage.__identifier__)
+        fti.behaviors = tuple(behaviors)
+        # invalidate schema cache
+        notify(SchemaInvalidatedEvent('collective.cover.content'))
+
+    def _disable_background_image_behavior(self):
+        fti = queryUtility(IDexterityFTI, name='collective.cover.content')
+        behaviors = list(fti.behaviors)
+        behaviors.remove(IBackgroundImage.__identifier__)
+        fti.behaviors = tuple(behaviors)
+        # invalidate schema cache
+        notify(SchemaInvalidatedEvent('collective.cover.content'))
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+
+        with api.env.adopt_roles(['Manager']):
+            self.folder = api.content.create(self.portal, 'Folder', 'test')
+
+        self._enable_background_image_behavior()
+        self.cover = api.content.create(
+            self.folder, 'collective.cover.content', 'c1')
+
+    def test_background_registration(self):
+        registration = queryUtility(IBehavior, name=IBackgroundImage.__identifier__)
+        self.assertIsNotNone(registration)
+
+    def test_background_in_cover(self):
+        fti = queryUtility(IDexterityFTI, name='collective.cover.content')
+        self.assertIn(IBackgroundImage.__identifier__, fti.behaviors)
+
+    def test_adapt_content(self):
+        background = IBackgroundImage(self.cover, None)
+        self.assertIsNotNone(background)
+
+        self._disable_background_image_behavior()
+        cover = api.content.create(
+            self.folder, 'collective.cover.content', 'c2')
+        background = IBackgroundImage(cover, None)
+        self.assertIsNone(background)